requirements.txt
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
psycopg2-binary==2.8.2
pytz==2019.1
